AHK/GBD: Improved late start detection
Now the original brightness is saved in an ini file and retrieved from there on every start. This solves the problem in that when the computer shuts-down in an abnormal way (like in a power outage), when the script starts again it incorretcly treats the modified value as the original value.