J/U/YSCH: Fixed the script breaking when YouTube is signed out
The new user blocking feature uses the comment's side-menu to add the new item, but YouTube only adds this side-menu if the user is signed in. This update fixes the problem, but currently you can't use the new button when signed out of YouTube. I may change this feature so that the button can be used either when signed in or out.